responsible_ai/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
boto3==1.38.16
botocore==1.38.16
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
concurrent-log-handler==0.9.26
contourpy==1.3.2
coverage==7.8.0
cycler==0.12.1
distro==1.9.0
filelock==3.18.0
Flask==3.1.1
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.1.1
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.1
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
json_repair==0.44.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
litellm==1.69.2
MarkupSafe==3.0.2
matplotlib==3.10.3
multidict==6.4.3
narwhals==1.39.0
networkx==3.4.2
numpy==2.2.5
openai==1.75.0
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pillow==11.2.1
plotly==6.0.1
pluggy==1.5.0
portalocker==3.1.1
propcache==0.3.1
protobuf==6.31.0
pyarrow==20.0.0
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
pyparsing==3.2.3
pytest==8.3.5
pytest-cov==6.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.24.0
s3transfer==0.12.0
scipy==1.15.3
seaborn==0.13.2
setuptools==78.1.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
statsmodels==0.14.4
streamlit==1.45.1
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
watchdog==6.0.0
Werkzeug==3.1.3
wheel==0.45.1
yarl==1.20.0
zipp==3.21.0
